Final Commit:
This is the final commit I have run the program and gradle is saying that it runs. I look a the tests and they come back with green lights if this isn't working or how you wanted it formatted pleas let me know either in an email at peteeri2@isu.edu or send me a pm on the discord I'm Eric Peterson.